gcs/requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pkg-resources==0.0.0
pyserial==3.4
Werkzeug==0.14.1
